Add Swagger/OpenAPI setup and refactor netinfo API into blueprint to enable automatic documentation. Also change netinfo endpoint.
